+++ src/main/scala/edu/cornell/cac/docker/api/Docker.scala
@@ -1,10 +1,10 @@
-package com.kolor.docker.api
+package edu.cornell.cac.docker.api
 
 import dispatch._
 import Defaults._
 import play.api.libs.iteratee.Enumerator
 import play.api.libs.json._
-import com.kolor.docker.api.entities.ContainerId
+import edu.cornell.cac.docker.api.entities.ContainerId
 import org.slf4j.LoggerFactory
 import scala.concurrent.Promise
 import play.api.libs.iteratee._
