+++ src/main/scala/edu/cornell/cac/docker/api/Exception.scala
@@ -1,6 +1,6 @@
-package com.kolor.docker.api
+package edu.cornell.cac.docker.api
 
-import com.kolor.docker.api.entities.ContainerId
+import edu.cornell.cac.docker.api.entities.ContainerId
 
 
 case class DockerResponseCode(code: Int, message: String) extends Exception(s"Docker error (Code $code): $message")
