+++ src/main/scala/edu/cornell/cac/docker/api/entities/ConfigReader.scala
@@ -1,4 +1,4 @@
-package com.kolor.docker.api.entities
+package edu.cornell.cac.docker.api.entities
 
 import scala.io.BufferedSource
