+++ src/main/scala/edu/cornell/cac/docker/api/entities/ContainerId.scala
@@ -1,6 +1,6 @@
-package com.kolor.docker.api.entities
+package edu.cornell.cac.docker.api.entities
 
-import com.kolor.docker.api._
+import edu.cornell.cac.docker.api._
 import play.api.libs.json._
 
 class ContainerId private[ContainerId] (val id: String) extends DockerEntity {
