+++ src/main/scala/edu/cornell/cac/docker/api/entities/DockerEntity.scala
@@ -1,11 +1,11 @@
 /**
  *
  */
-package com.kolor.docker.api.entities
+package edu.cornell.cac.docker.api.entities
 
 import org.joda.time.DateTime
 import play.api.libs.json._
-import com.kolor.docker.api._
+import edu.cornell.cac.docker.api._
 
 
 trait DockerEntity {
