+++ src/main/scala/edu/cornell/cac/docker/api/entities/DockerVolume.scala
@@ -1,6 +1,6 @@
-package com.kolor.docker.api.entities
+package edu.cornell.cac.docker.api.entities
 
-import com.kolor.docker.api._
+import edu.cornell.cac.docker.api._
 import play.api.libs.json._
 
 trait DockerVolume extends DockerEntity {
