+++ src/main/scala/edu/cornell/cac/docker/dsl/package.scala
@@ -1,4 +1,4 @@
-package com.kolor.docker
+package edu.cornell.cac.docker
 
 package object dsl {
